requirements.txt
spacy>=3.0
vosk>=0.3.45
soundfile>=0.10.3
numpy>=1.21

# LLM Providers (install the one you need)
groq>=0.4.0              # FREE, fast - recommended
openai>=1.0.0            # For OpenAI/Together.ai
python-dotenv>=1.0.0     # For .env file support
requests>=2.28.0         # For Ollama local
huggingface_hub>=0.20.0  # For HuggingFace API

# Get free API keys:
# Groq: https://console.groq.com/keys
# HuggingFace: https://huggingface.co/settings/tokens
